+++ test/01-manipulating-databases.js
@@ -9,7 +9,7 @@ describe('Manipulating databases', () => {
   let db
   beforeEach(() => {
     db = new Database({
-      url: (process.env.TEST_ARANGODB_URL || 'http://root:@localhost:8529'),
+      url: (process.env.TEST_ARANGODB_URL || 'vstream://root:@localhost:8529'),
       arangoVersion: Number(process.env.ARANGO_VERSION || 30000)
     })
   })
