+++ test/05-aql-queries.js
@@ -8,7 +8,7 @@ describe('AQL queries', () => {
   let db
   before((done) => {
     db = new Database({
-      url: (process.env.TEST_ARANGODB_URL || 'http://root:@localhost:8529'),
+      url: (process.env.TEST_ARANGODB_URL || 'vstream://root:@localhost:8529'),
       arangoVersion: Number(process.env.ARANGO_VERSION || 30000)
     })
     db.createDatabase(name)
