+++ test/08-cursors.js
@@ -10,7 +10,7 @@ describe('Cursor API', () => {
   let cursor
   before(() => {
     db = new Database({
-      url: (process.env.TEST_ARANGODB_URL || 'http://root:@localhost:8529'),
+      url: (process.env.TEST_ARANGODB_URL || 'vstream://root:@localhost:8529'),
       arangoVersion: Number(process.env.ARANGO_VERSION || 30000)
     })
   })
