+++ test/14-document-collections.js
@@ -11,7 +11,7 @@ describe('DocumentCollection API', () => {
   let collection
   before((done) => {
     db = new Database({
-      url: (process.env.TEST_ARANGODB_URL || 'http://root:@localhost:8529'),
+      url: (process.env.TEST_ARANGODB_URL || 'vstream://root:@localhost:8529'),
       arangoVersion: ARANGO_VERSION
     })
     db.createDatabase(name)
