+++ test/16-graphs.js
@@ -32,7 +32,7 @@ describe('Graph API', () => {
   let name = `testdb_${Date.now()}`
   before((done) => {
     db = new Database({
-      url: (process.env.TEST_ARANGODB_URL || 'http://root:@localhost:8529'),
+      url: (process.env.TEST_ARANGODB_URL || 'vstream://root:@localhost:8529'),
       arangoVersion: Number(process.env.ARANGO_VERSION || 30000)
     })
     db.createDatabase(name)
