+++ test/17-graph-vertices.js
@@ -36,7 +36,7 @@ describe('Manipulating graph vertices', () => {
   let collectionNames
   before((done) => {
     db = new Database({
-      url: (process.env.TEST_ARANGODB_URL || 'http://root:@localhost:8529'),
+      url: (process.env.TEST_ARANGODB_URL || 'vstream://root:@localhost:8529'),
       arangoVersion: Number(process.env.ARANGO_VERSION || 30000)
     })
     db.createDatabase(name)
